pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pybspf"
version = "0.1.0"
description = "pybspf: B-spline plus spectral/Fourier operators for 1D sampled data."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "Proprietary"}
authors = [
  {name = "pybspf contributors"},
]
dependencies = [
  "numpy>=1.24",
  "scipy>=1.10",
]

[project.optional-dependencies]
gpu = [
  "cupy-cuda12x",
]
dev = [
  "pytest>=8.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["pybspf*"]
